requirements.txt
configparser==5.0.0
pytest==6.0.1
selenium==3.141.0
webdriver-manager==3.2.2
suds-py3== 1.4.1.0